requirements.txt
mysql-connector-python==8.0.27
pandas==1.3.5
numpy==1.22.1
protobuf==3.19.3
python_dateutil==2.8.2
pytz==2021.3
setuptools==60.2.0
six==1.16.0
wheel==0.37.1